# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "ATWT API for Testing",
        "description": "This is an example api for testing purpose. automatedtestingwithtuyen.com",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "contact": {
            "name": "Tuyen Nguyen",
            "url": "http://automatedtestingwithtuyen.com",
            "email": "trongtuyen96@gmail.com"
        }
    },
    "host": "atwt-api.herokuapp.com",
    "basePath": "/v1",
...
    "schemes": [
        "https"
    ],
    "paths": {
...
        "/zoo": {
            "get": {
                "tags": [
                    "zoo"
                ],
                "summary": "Get zoos with conditions",
                "description": "",
                "operationId": "getZoos",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "name",
                        "description": "Name of zoo",
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "country",
                        "description": "Country of zoo",
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "license",
                        "description": "License of zoo",
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "year",
                        "description": "Year of establishment of zoo (greater than or equal)",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfull operation",
                        "schema": {
                            "type": "array",
                            "$ref": "#/definitions/Zoo"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "zoo"
                ],
                "summary": "Add a new zoo",
                "description": "",
                "operationId": "postZoo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Zoo object that need to add",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Zoo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfull operation"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/zoo/{zooId}": {
            "get": {
                "tags": [
                    "zoo"
                ],
                "summary": "Find zoo by Id",
                "description": "Returns a single zoo",
                "operationId": "getZooById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "zooId",
                        "in": "path",
                        "description": "Id of zoo to return",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Zoo"
                        }
                    },
                    "400": {
                        "description": "Invalid Id supplied"
                    },
                    "404": {
                        "description": "Zoo not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "zoo"
                ],
                "summary": "Update an existing zoo",
                "description": "",
                "operationId": "putZoo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "zooId",
                        "in": "path",
                        "description": "Id of zoo to update",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Zoo object that needs to be updated",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Zoo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Update successfully"
                    },
                    "400": {
                        "description": "Invalid Id supplied"
                    },
                    "404": {
                        "description": "Zoo not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            },
            "delete": {
                "tags": [
                    "zoo"
                ],
                "summary": "Delete a zoo",
                "description": "",
                "operationId": "deleteZoo",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "zooId",
                        "in": "path",
                        "description": "Zoo Id to delete",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid Id supplied"
                    },
                    "404": {
                        "description": "Zoo not found"
                    }
                },
                "deprecated": true
            }
        },
        "/zoo/{zooId}/animals": {
            "get": {
                "tags": [
                    "zoo"
                ],
                "summary": "Find all animals in the zoo by Id",
                "description": "Returns animals in the zoo",
                "operationId": "getAllAnimalsOfZooById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "zooId",
                        "in": "path",
                        "description": "Id of zoo to return",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "$ref": "#/definitions/Animal"
                        }
                    },
                    "400": {
                        "description": "Invalid Id supplied"
                    },
                    "404": {
                        "description": "Zoo not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Zoo": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "license": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            },
            "xml": {
                "name": "Zoo"
            }
        },
        "Animal": {
            "type": "object",
            "required": [
                "zooID"
            ],
            "properties": {
                "species": {
                    "type": "string"
                },
                "lifespan": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "weight": {
                    "type": "integer"
                },
                "zooID": {
                    "type": "string"
                }
            },
            "xml": {
                "name": "Animal"
            }
        }
    }
}